requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-settings
motor
python-dotenv
passlib[bcrypt]
python-jose[cryptography]